backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
psycopg2-binary>=2.9
python-decouple>=3.8
pillow>=10.0
celery>=5.3
redis>=5.0
pytest>=7.4
pytest-django>=4.7
factory-boy>=3.3
